add checker texture to the world

+++ rt_next_week/cuda/sphere.h
@@ -31,8 +31,8 @@ protected:
     //     <1 0 0> yields <0.50 0.50>       <-1  0  0> yields <0.00 0.50>
     //     <0 1 0> yields <0.50 1.00>       < 0 -1  0> yields <0.50 0.00>
     //     <0 0 1> yields <0.25 0.50>       < 0  0 -1> yields <0.75 0.50>
-    float theta = acos(-p.y());
-    float phi = atan2(-p.z(), p.x()) + pi;
+    float theta = acosf(-p.y());
+    float phi = atan2f(-p.z(), p.x()) + pi;
 
     u = phi / (2.0f * pi);
     v = theta / pi;
